Simplify code now that we natively support reading/writing locations in a SymbolKey.
